resume_evaluator/config.py
"""Configuration management for the resume evaluator."""

import os
from typing import Optional

from dotenv import load_dotenv

# Load environment variables from .env file
load_dotenv()


def get_openai_api_key() -> str:
    """
    Get OpenAI API key from environment variable.

    Returns:
        The OpenAI API key.

    Raises:
        ValueError: If OPENAI_API_KEY is not set in the environment.
    """
    api_key = os.getenv("OPENAI_API_KEY")
    if not api_key:
        raise ValueError(
            "OPENAI_API_KEY environment variable is not set. "
            "Please set it in your .env file or environment."
        )
    return api_key
