requirements.txt
# This file was autogenerated by uv via the following command:
#    uv pip compile pyproject.toml -o requirements.txt
altair==6.0.0
    # via streamlit
annotated-doc==0.0.4
    # via fastapi
annotated-types==0.7.0
    # via pydantic
anyio==4.12.0
    # via
    #   httpx
    #   openai
    #   starlette
    #   watchfiles
attrs==25.4.0
    # via
    #   jsonschema
    #   referencing
blinker==1.9.0
    # via streamlit
cachetools==6.2.4
    # via streamlit
certifi==2025.11.12
    # via
    #   httpcore
    #   httpx
    #   requests
charset-normalizer==3.4.4
    # via requests
click==8.3.1
    # via
    #   streamlit
    #   uvicorn
distro==1.9.0
    # via openai
fastapi==0.128.0
    # via resume-evaluator (pyproject.toml)
gitdb==4.0.12
    # via gitpython
gitpython==3.1.45
    # via streamlit
h11==0.16.0
    # via
    #   httpcore
    #   uvicorn
httpcore==1.0.9
    # via httpx
httptools==0.7.1
    # via uvicorn
httpx==0.28.1
    # via
    #   langgraph-sdk
    #   langsmith
    #   openai
idna==3.11
    # via
    #   anyio
    #   httpx
    #   requests
jinja2==3.1.6
    # via
    #   altair
    #   pydeck
jiter==0.12.0
    # via openai
jsonpatch==1.33
    # via langchain-core
jsonpointer==3.0.0
    # via jsonpatch
jsonschema==4.25.1
    # via altair
jsonschema-specifications==2025.9.1
    # via jsonschema
langchain-core==1.2.5
    # via
    #   langgraph
    #   langgraph-checkpoint
    #   langgraph-prebuilt
langgraph==1.0.5
    # via resume-evaluator (pyproject.toml)
langgraph-checkpoint==3.0.1
    # via
    #   langgraph
    #   langgraph-prebuilt
langgraph-prebuilt==1.0.5
    # via langgraph
langgraph-sdk==0.3.1
    # via langgraph
langsmith==0.5.2
    # via langchain-core
markupsafe==3.0.3
    # via jinja2
narwhals==2.14.0
    # via altair
numpy==2.4.0
    # via
    #   pandas
    #   pydeck
    #   streamlit
openai==2.14.0
    # via resume-evaluator (pyproject.toml)
orjson==3.11.5
    # via
    #   langgraph-sdk
    #   langsmith
ormsgpack==1.12.1
    # via langgraph-checkpoint
packaging==25.0
    # via
    #   altair
    #   langchain-core
    #   langsmith
    #   streamlit
pandas==2.3.3
    # via streamlit
pillow==12.0.0
    # via streamlit
protobuf==6.33.2
    # via streamlit
pyarrow==22.0.0
    # via streamlit
pydantic==2.12.5
    # via
    #   resume-evaluator (pyproject.toml)
    #   fastapi
    #   langchain-core
    #   langgraph
    #   langsmith
    #   openai
pydantic-core==2.41.5
    # via pydantic
pydeck==0.9.1
    # via streamlit
pypdf==6.5.0
    # via resume-evaluator (pyproject.toml)
python-dateutil==2.9.0.post0
    # via pandas
python-dotenv==1.2.1
    # via
    #   resume-evaluator (pyproject.toml)
    #   uvicorn
python-multipart==0.0.21
    # via resume-evaluator (pyproject.toml)
pytz==2025.2
    # via pandas
pyyaml==6.0.3
    # via
    #   langchain-core
    #   uvicorn
referencing==0.37.0
    # via
    #   jsonschema
    #   jsonschema-specifications
requests==2.32.5
    # via
    #   langsmith
    #   requests-toolbelt
    #   streamlit
requests-toolbelt==1.0.0
    # via langsmith
rpds-py==0.30.0
    # via
    #   jsonschema
    #   referencing
six==1.17.0
    # via python-dateutil
smmap==5.0.2
    # via gitdb
sniffio==1.3.1
    # via openai
starlette==0.50.0
    # via fastapi
streamlit==1.52.2
    # via resume-evaluator (pyproject.toml)
tenacity==9.1.2
    # via
    #   resume-evaluator (pyproject.toml)
    #   langchain-core
    #   streamlit
toml==0.10.2
    # via streamlit
tornado==6.5.4
    # via streamlit
tqdm==4.67.1
    # via
    #   resume-evaluator (pyproject.toml)
    #   openai
typing-extensions==4.15.0
    # via
    #   altair
    #   anyio
    #   fastapi
    #   langchain-core
    #   openai
    #   pydantic
    #   pydantic-core
    #   referencing
    #   starlette
    #   streamlit
    #   typing-inspection
typing-inspection==0.4.2
    # via pydantic
tzdata==2025.3
    # via pandas
urllib3==2.6.2
    # via requests
uuid-utils==0.12.0
    # via
    #   langchain-core
    #   langsmith
uvicorn==0.40.0
    # via resume-evaluator (pyproject.toml)
uvloop==0.22.1
    # via uvicorn
watchfiles==1.1.1
    # via uvicorn
websockets==15.0.1
    # via uvicorn
xxhash==3.6.0
    # via langgraph
zstandard==0.25.0
    # via langsmith
